Revert "Try to implement test cases for artifact store (#3154)"
This reverts commit 3c553bb57bdffb306a272d60fd4158ed8e72df02.
