+++ backend/common/src/main/java/ai/verta/modeldb/common/config/ArtifactStoreConfig.java
@@ -62,14 +62,6 @@ public class ArtifactStoreConfig {
     }
   }
 
-  public void setHost(String host) {
-    this.host = host;
-  }
-
-  public void setNFS(NFSConfig nfs) {
-    this.NFS = nfs;
-  }
-
   public String storeTypePathPrefix() {
     switch (getArtifactStoreType()) {
       case "S3":
