+++ backend/common/src/main/java/ai/verta/modeldb/common/config/Config.java
@@ -239,10 +239,6 @@ public abstract class Config {
     return openTelemetry;
   }
 
-  public void setArtifactStoreConfig(ArtifactStoreConfig artifactStoreConfig) {
-    this.artifactStoreConfig = artifactStoreConfig;
-  }
-
   public void setSpringServer(SpringServerConfig springServer) {
     this.springServer = springServer;
   }
