+++ backend/common/src/main/java/ai/verta/modeldb/common/configuration/ArtifactStoreInitBeans.java
@@ -45,8 +45,7 @@ public class ArtifactStoreInitBeans {
 
       if (artifactStoreConfig.getArtifactStoreType().equals("NFS")
           && artifactStoreConfig.getNFS() != null
-          && artifactStoreConfig.getNFS().getArtifactEndpoint() != null
-          && artifactStoreConfig.getArtifactEndpoint() == null) {
+          && artifactStoreConfig.getNFS().getArtifactEndpoint() != null) {
         System.getProperties()
             .put(
                 "artifactEndpoint.storeArtifact",
