+++ backend/src/main/java/ai/verta/modeldb/config/TestConfig.java
@@ -2,15 +2,11 @@ package ai.verta.modeldb.config;
 
 import ai.verta.modeldb.ModelDBConstants;
 import ai.verta.modeldb.common.CommonMessages;
-import ai.verta.modeldb.common.CommonUtils;
 import ai.verta.modeldb.common.config.InvalidConfigException;
 import ai.verta.modeldb.common.config.ServiceUserConfig;
 import ai.verta.modeldb.common.exceptions.InternalErrorException;
-import ai.verta.modeldb.common.exceptions.ModelDBException;
 import ai.verta.modeldb.common.futures.FutureJdbi;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.io.FileInputStream;
-import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.AccessLevel;
@@ -18,9 +14,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
-import org.yaml.snakeyaml.Yaml;
-import org.yaml.snakeyaml.constructor.Constructor;
-import org.yaml.snakeyaml.introspector.BeanAccess;
 
 @Data
 @AllArgsConstructor
@@ -37,29 +30,6 @@ public class TestConfig extends MDBConfig {
       config.validate();
     }
     return config;
-  }
-
-  public static TestConfig getInstance(String path) throws InternalErrorException {
-    if (config == null) {
-      config = readConfig(TestConfig.class, path);
-      config.validate();
-    }
-    return config;
-  }
-
-  private static <T> T readConfig(Class<T> configType, String configFile)
-      throws InternalErrorException {
-    try {
-      var yaml = new Yaml(new Constructor(configType));
-      configFile = CommonUtils.appendOptionalTelepresencePath(configFile);
-      InputStream inputStream = new FileInputStream(configFile);
-      yaml.setBeanAccess(BeanAccess.FIELD);
-      return yaml.loadAs(inputStream, configType);
-    } catch (ModelDBException | NullPointerException ex) {
-      throw ex;
-    } catch (Exception ex) {
-      throw new InternalErrorException(ex.getMessage());
-    }
   }
 
   @Override
