+++ backend/src/main/java/ai/verta/modeldb/configuration/AppConfigBeans.java
@@ -15,7 +15,6 @@ import ai.verta.modeldb.common.configuration.AppContext;
 import ai.verta.modeldb.common.configuration.EnabledMigration;
 import ai.verta.modeldb.common.configuration.RunLiquibaseSeparately;
 import ai.verta.modeldb.common.configuration.RunLiquibaseSeparately.RunLiquibaseWithMainService;
-import ai.verta.modeldb.common.connections.UAC;
 import ai.verta.modeldb.common.exceptions.ExceptionInterceptor;
 import ai.verta.modeldb.common.futures.FutureUtil;
 import ai.verta.modeldb.common.interceptors.MetadataForwarder;
@@ -86,20 +85,12 @@ public class AppConfigBeans {
   public MDBConfig config() {
     var config = MDBConfig.getInstance();
     App.getInstance().mdbConfig = config;
-    initializeSystemProperties(config);
-    return config;
-  }
 
-  private static void initializeSystemProperties(MDBConfig config) {
     // Configure spring HTTP server
-    var webPort = config.getSpringServer().getPort();
-    if (webPort == 0) {
-      var grpcServerConfig = config.getGrpcServer();
-      var grpcServerPort = grpcServerConfig.getPort();
-      webPort = grpcServerPort + 1;
-    }
-    System.getProperties().put("server.port", webPort);
-    LOGGER.info("Configuring spring HTTP traffic on port: {}", webPort);
+    LOGGER.info("Configuring spring HTTP traffic on port: {}", config.getSpringServer().getPort());
+    System.getProperties().put("server.port", config.getSpringServer().getPort());
+
+    return config;
   }
 
   @Bean
@@ -114,15 +105,10 @@ public class AppConfigBeans {
   }
 
   @Bean
-  UAC uac(Config config) {
-    return UAC.FromConfig(config);
-  }
-
-  @Bean
-  ServiceSet serviceSet(MDBConfig config, ArtifactStoreService artifactStoreService, UAC uac)
+  ServiceSet serviceSet(MDBConfig config, ArtifactStoreService artifactStoreService)
       throws IOException {
     // Initialize services that we depend on
-    return ServiceSet.fromConfig(config, artifactStoreService, uac);
+    return ServiceSet.fromConfig(config, artifactStoreService);
   }
 
   @Bean
@@ -212,6 +198,10 @@ public class AppConfigBeans {
         up.inc();
         LOGGER.info("Current PID: {}", ProcessHandle.current().pid());
         LOGGER.info("Backend server started listening on {}", config.getGrpcServer().getPort());
+
+        // ----------- Don't exit the main thread. Wait until server is terminated -----------
+        server.awaitTermination();
+        up.dec();
       } catch (Exception ex) {
         CommonUtils.printStackTrace(LOGGER, ex);
         appContext.initiateShutdown(0);
