+++ backend/src/main/java/ai/verta/modeldb/experimentRun/FutureExperimentRunDAO.java
@@ -516,9 +516,8 @@ public class FutureExperimentRunDAO {
             handle ->
                 handle
                     .createQuery(
-                        "SELECT project_id FROM experiment_run WHERE id IN (<ids>) AND deleted=:deleted")
+                        "SELECT project_id FROM experiment_run WHERE id IN (<ids>) AND deleted=0")
                     .bindList("ids", finalRunIds)
-                    .bind("deleted", false)
                     .mapTo(String.class)
                     .list());
