+++ backend/src/main/resources/application.properties
@@ -1,2 +1,1 @@
 spring.main.banner-mode=off
-spring.main.allow-bean-definition-overriding=true
