--- backend/src/test/java/ai/verta/modeldb/ArtifactStore/nfs-config-test-h2.yaml
@@ -1,71 +1,0 @@
-#ModelDB Properties
-grpcServer:
-  port: 99999
-  requestTimeout: 30
-  metrics_port: 99999
-
-springServer:
-  port: 99999
-  shutdownTimeout: 30
-service_user:
-  email: registry_admin@verta.ai
-  devKey: abc111
-
-database:
-  DBType: relational
-  timeout: 4
-  liquibaseLockThreshold: 60
-  RdbConfiguration:
-    RdbDatabaseName: "artifactStoreTestDB"
-    RdbDriver: org.h2.Driver
-    RdbDialect: org.hibernate.dialect.H2Dialect
-    RdbUrl: "jdbc:h2:mem:artifactStoreTestDB"
-    RdbUsername: sa
-    RdbPassword: password
-    sslMode: DISABLED
-
-testUsers:
-  primaryUser:
-    email: registry_test1@verta.ai
-    devKey: abc123
-  secondaryUser:
-    email: registry_test2@verta.ai
-    devKey: abc456
-
-artifactStoreConfig:
-  artifactStoreType: NFS
-  enabled: true
-  protocol: http
-  host: localhost:99999
-  pickArtifactStoreHostFromConfig: true
-  artifactEndpoint:
-    getArtifact: "/v1/artifact/getArtifact"
-    storeArtifact: "/v1/artifact/storeArtifact"
-  S3:
-    cloudAccessKey:
-    cloudSecretKey:
-    cloudBucketName:
-    awsRegion:
-    s3presignedURLEnabled:
-    minioEndpoint:
-  NFS:
-    nfsRootPath: ./
-    nfsServerHost: localhost
-    nfsUrlProtocol: http
-    nfsPathPrefix: test_prefix
-    artifactEndpoint:
-      getArtifact: "/v1/artifact/getArtifact"
-      storeArtifact: "/v1/artifact/storeArtifact"
-
-authService:
-  host: localhost
-  port: 50051
-
-telemetry:
- opt_out: true
-
-migrations:
-
-cron_job:
-
-populateConnectionsBasedOnPrivileges: true
